+++ KiloMart.Commands/Services/ProductOfferService.cs
@@ -11,7 +11,7 @@ namespace KiloMart.Commands.Services
         public decimal Price { get; set; }
         public decimal OffPercentage { get; set; }
         //public DateTime FromDate { get; set; }
-        public float Quantity { get; set; }
+        public decimal Quantity { get; set; }
 
         public (bool Success, string[] Errors) Validate()
         {
@@ -40,7 +40,7 @@ namespace KiloMart.Commands.Services
     {
         public int Id { get; set; }
         public decimal? OffPercentage { get; set; }
-        public float? Quantity { get; set; }
+        public decimal? Quantity { get; set; }
 
         public (bool Success, string[] Errors) Validate()
         {
