+++ KiloMart.DataAccess.EFCore/Models/Location.cs
@@ -7,9 +7,9 @@ public partial class Location
 {
     public int Id { get; set; }
 
-    public double Longitude { get; set; }
+    public decimal Longitude { get; set; }
 
-    public double Latitude { get; set; }
+    public decimal Latitude { get; set; }
 
     public string Name { get; set; } = null!;
